# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Librar API",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/reviews": {
            "get": {
                "tags": [
                    "Reviews"
                ],
                "summary": "Получение отзывов",
                "operationId": "getReviews",
                "parameters": [
                    {
                        "$ref": "#/components/parameters/SearchQuery"
                    },
                    {
                        "$ref": "#/components/parameters/IdQuery"
                    },
                    {
                        "$ref": "#/components/parameters/PerPageQuery"
                    }
                ],
                "responses": {
                    "200": {
                        "$ref": "#/components/responses/ReviewsResponse"
                    },
                    "404": {
                        "$ref": "#/components/responses/NotFound"
                    }
                }
            }
        },
        "/api/admin/reviews": {
            "get": {
                "tags": [
                    "Reviews"
                ],
                "summary": "Получение отзывов для админа",
                "operationId": "getAdminReviews",
                "parameters": [
                    {
                        "$ref": "#/components/parameters/SearchQuery"
                    },
                    {
                        "$ref": "#/components/parameters/IdQuery"
                    },
                    {
                        "$ref": "#/components/parameters/BookIdQuery"
                    },
                    {
                        "$ref": "#/components/parameters/EmailQuery"
                    },
                    {
                        "$ref": "#/components/parameters/PerPageQuery"
                    }
                ],
                "responses": {
                    "200": {
                        "$ref": "#/components/responses/ReviewsResponse"
                    },
                    "404": {
                        "$ref": "#/components/responses/NotFound"
                    }
                }
            }
        },
        "/api/books/{bookId}/reviews": {
            "get": {
                "tags": [
                    "Reviews"
                ],
                "summary": "Получение отзывов книги",
                "operationId": "getBooksReviews",
                "parameters": [
                    {
                        "$ref": "#/components/parameters/BookIdInPath"
                    }
                ],
                "responses": {
                    "200": {
                        "$ref": "#/components/responses/ReviewsSummaryResponse"
                    },
                    "404": {
                        "$ref": "#/components/responses/NotFound"
                    }
                }
            },
            "post": {
                "tags": [
                    "Reviews"
                ],
                "summary": "Создать отзыв",
                "operationId": "createReview",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/StoreReviewRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "$ref": "#/components/responses/ReviewResponse"
                    }
                }
            }
        },
        "/api/reviews/user/{userId}": {
            "get": {
                "tags": [
                    "Reviews"
                ],
                "summary": "Получение отзывов пользователя",
                "operationId": "getUserReviews",
                "parameters": [
                    {
                        "$ref": "#/components/parameters/UserIdInPath"
                    }
                ],
                "responses": {
                    "200": {
                        "$ref": "#/components/responses/ReviewsSummaryResponse"
                    },
                    "404": {
                        "$ref": "#/components/responses/NotFound"
                    }
                }
            }
        },
        "/api/reviews/{id}": {
            "put": {
                "tags": [
                    "Reviews"
                ],
                "summary": "Изменить отзыв",
                "operationId": "updateReview",
                "parameters": [
                    {
                        "$ref": "#/components/parameters/IdInPath"
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/StoreReviewRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "$ref": "#/components/responses/ReviewResponse"
                    }
                }
            },
            "post": {
                "tags": [
                    "Reviews"
                ],
                "summary": "Восстановить отзыв",
                "operationId": "restoreReview",
                "parameters": [
                    {
                        "$ref": "#/components/parameters/IdInPath"
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Восстановлено"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Reviews"
                ],
                "summary": "Удалить отзыв",
                "operationId": "deleteReview",
                "parameters": [
                    {
                        "$ref": "#/components/parameters/IdInPath"
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Удалено"
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
...
            "StoreReviewRequest": {
                "required": [
                    "rating"
                ],
                "properties": {
                    "text": {
                        "type": "string",
                        "example": "Текст отзыва"
                    },
                    "rating": {
                        "type": "integer",
                        "example": "Оценка"
                    }
                },
                "type": "object"
            },
...
            "AuthorSummaryResource": {
                "properties": {
                    "id": {
                        "type": "integer"
                    },
                    "name": {
                        "type": "string"
                    },
                    "slug": {
                        "type": "string"
                    }
                },
                "type": "object"
            },
...
            "BookSummaryResource": {
                "properties": {
                    "id": {
                        "type": "integer"
                    },
                    "title": {
                        "type": "string"
                    },
                    "slug": {
                        "type": "string"
                    },
                    "image": {
                        "type": "string"
                    },
                    "author": {
                        "$ref": "#/components/schemas/AuthorSummaryResource"
                    },
                    "isAvailable": {
                        "type": "boolean"
                    }
                },
                "type": "object"
            },
...
            "ReviewCollection": {
                "properties": {
                    "data": {
                        "description": "Список отзывов",
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/ReviewResource"
                        }
                    }
                },
                "type": "object"
            },
            "ReviewResource": {
                "properties": {
                    "id": {
                        "type": "integer"
                    },
                    "rating": {
                        "type": "integer"
                    },
                    "text": {
                        "type": "string"
                    },
                    "book": {
                        "$ref": "#/components/schemas/BookSummaryResource"
                    },
                    "user": {
                        "$ref": "#/components/schemas/UserResource"
                    },
                    "createdAt": {
                        "type": "integer"
                    },
                    "isDeleted": {
                        "type": "boolean"
                    }
                },
                "type": "object"
            },
            "ReviewSummaryCollection": {
                "properties": {
                    "data": {
                        "description": "Список отзывов",
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/ReviewSummaryResource"
                        }
                    }
                },
                "type": "object"
            },
            "ReviewSummaryResource": {
                "properties": {
                    "id": {
                        "type": "integer"
                    },
                    "rating": {
                        "type": "integer"
                    },
                    "text": {
                        "type": "string"
                    },
                    "book": {
                        "$ref": "#/components/schemas/BookSummaryResource"
                    },
                    "user": {
                        "$ref": "#/components/schemas/UserResource"
                    },
                    "createdAt": {
                        "type": "integer"
                    }
                },
                "type": "object"
            },
...
            "UserResource": {
                "properties": {
                    "id": {
                        "type": "integer"
                    },
                    "name": {
                        "type": "string"
                    },
                    "email": {
                        "type": "string"
                    },
                    "role": {
                        "type": "string"
                    },
                    "notifications": {
                        "type": "integer"
                    },
                    "isVerified": {
                        "type": "boolean"
                    },
                    "isDeleted": {
                        "type": "boolean"
                    }
                },
                "type": "object"
            }
        },
        "responses": {
...
            "NotFound": {
                "description": "Не найдено"
            },
...
            "ReviewResponse": {
                "description": "Отзыв",
                "content": {
                    "application/json": {
                        "schema": {
                            "$ref": "#/components/schemas/ReviewResource"
                        }
                    }
                }
            },
            "ReviewsResponse": {
                "description": "Отзывы",
                "content": {
                    "application/json": {
                        "schema": {
                            "$ref": "#/components/schemas/ReviewCollection"
                        }
                    }
                }
            },
            "ReviewsSummaryResponse": {
                "description": "Отзывы",
                "content": {
                    "application/json": {
                        "schema": {
                            "$ref": "#/components/schemas/ReviewSummaryCollection"
                        }
                    }
                }
            }
        },
        "parameters": {
            "BookIdInPath": {
                "name": "bookId",
                "in": "path",
                "description": "ID книги",
                "required": true,
                "schema": {
                    "type": "integer"
                }
            },
            "BookIdQuery": {
                "name": "bookId",
                "in": "query",
                "description": "bookId ресурса",
                "schema": {
                    "type": "integer"
                }
            },
...
            "EmailQuery": {
                "name": "email",
                "in": "query",
                "description": "email ресурса",
                "schema": {
                    "type": "string"
                }
            },
...
            "IdInPath": {
                "name": "id",
                "in": "path",
                "description": "ID ресурса",
                "required": true,
                "schema": {
                    "type": "integer"
                }
            },
            "IdQuery": {
                "name": "id",
                "in": "query",
                "description": "id ресурса",
                "schema": {
                    "type": "integer"
                }
            },
...
            "PerPageQuery": {
                "name": "perPage",
                "in": "query",
                "description": "Сколько страниц",
                "schema": {
                    "type": "integer",
                    "default": 10
                }
            },
...
            "SearchQuery": {
                "name": "q",
                "in": "query",
                "description": "Поиск ресурса",
                "schema": {
                    "type": "string"
                }
            },
...
            "UserIdInPath": {
                "name": "userId",
                "in": "path",
                "description": "ID пользователя",
                "required": true,
                "schema": {
                    "type": "integer"
                }
            },
...
        }
    },
...
}